+++ examples/extras/szengine/add_record_with_info.py
@@ -39,7 +39,7 @@ RECORD_ID = "1001"
 
 try:
     sz_abstract_factory = SzAbstractFactory(**FACTORY_PARAMETERS)
-    sz_engine = sz_abstract_factory.create_sz_engine()
+    sz_engine = sz_abstract_factory.create_engine()
     RESULT = sz_engine.add_record(DATA_SOURCE_CODE, RECORD_ID, RECORD_DEFINITION, SzEngineFlags.SZ_WITH_INFO)
     print(f"\nFile {__file__}:\n{RESULT}\n")
 except SzError as err: