+++ examples/szengine/preprocess_record.py
@@ -38,7 +38,7 @@ RECORD_DEFINITION = (
 
 try:
     sz_abstract_factory = SzAbstractFactory(**FACTORY_PARAMETERS)
-    sz_engine = sz_abstract_factory.create_sz_engine()
+    sz_engine = sz_abstract_factory.create_engine()
     RESULT = sz_engine.preprocess_record(RECORD_DEFINITION, FLAGS)
     print(f"\nFile {__file__}:\n{RESULT}\n")
 except SzError as err: