+++ src/senzing/szabstractfactory.py
@@ -55,7 +55,7 @@ class SzAbstractFactoryParameters(TypedDict, total=False):
 
 class SzAbstractFactory(SzAbstractFactoryAbstract):
     """
-    SzAbstractFactory module is a factory pattern for accessing Senzing over gRPC.
+    SzAbstractFactory module is a factory pattern for accessing Senzing.
     """
 
     # -------------------------------------------------------------------------
@@ -68,7 +68,6 @@ class SzAbstractFactory(SzAbstractFactoryAbstract):
         settings: Union[str, Dict[Any, Any]] = "",
         config_id: int = 0,
         verbose_logging: int = 0,
-        **kwargs: Any,
     ) -> None:
         """
         Constructor
@@ -87,7 +86,7 @@ class SzAbstractFactory(SzAbstractFactoryAbstract):
 
     def __enter__(
         self,
-    ) -> Any:  # TODO: Replace "Any" with "Self" once python 3.11 is lowest supported python version.
+    ) -> Any:
         """Context Manager method."""
         return self
 
@@ -108,8 +107,7 @@ class SzAbstractFactory(SzAbstractFactoryAbstract):
     # SzAbstractFactory methods
     # -------------------------------------------------------------------------
 
-    def create_sz_config(self, **kwargs: Any) -> SzConfigAbstract:
-        _ = kwargs
+    def create_config(self) -> SzConfigAbstract:
         result = SzConfig()
         if not self.is_szconfig_initialized:
             result._initialize(  # pylint: disable=W0212
@@ -120,8 +118,7 @@ class SzAbstractFactory(SzAbstractFactoryAbstract):
             self.is_szconfig_initialized = True
         return result
 
-    def create_sz_configmanager(self, **kwargs: Any) -> SzConfigManagerAbstract:
-        _ = kwargs
+    def create_configmanager(self) -> SzConfigManagerAbstract:
         result = SzConfigManager()
         if not self.is_szconfigmanager_initialized:
             result._initialize(  # pylint: disable=W0212
@@ -132,8 +129,7 @@ class SzAbstractFactory(SzAbstractFactoryAbstract):
             self.is_szconfigmanager_initialized = True
         return result
 
-    def create_sz_diagnostic(self, **kwargs: Any) -> SzDiagnosticAbstract:
-        _ = kwargs
+    def create_diagnostic(self) -> SzDiagnosticAbstract:
         result = SzDiagnostic()
         if not self.is_szdiagnostic_initialized:
             result._initialize(  # pylint: disable=W0212
@@ -145,8 +141,7 @@ class SzAbstractFactory(SzAbstractFactoryAbstract):
             self.is_szdiagnostic_initialized = True
         return result
 
-    def create_sz_engine(self, **kwargs: Any) -> SzEngineAbstract:
-        _ = kwargs
+    def create_engine(self) -> SzEngineAbstract:
         # TODO: Determine if atomic operation is needed.
         result = SzEngine()
         if not self.is_szengine_initialized:
@@ -159,8 +154,7 @@ class SzAbstractFactory(SzAbstractFactoryAbstract):
             self.is_szengine_initialized = True
         return result
 
-    def create_sz_product(self, **kwargs: Any) -> SzProductAbstract:
-        _ = kwargs
+    def create_product(self) -> SzProductAbstract:
         result = SzProduct()
         if not self.is_szproduct_initialized:
             result._initialize(  # pylint: disable=W0212
@@ -171,7 +165,7 @@ class SzAbstractFactory(SzAbstractFactoryAbstract):
             self.is_szproduct_initialized = True
         return result
 
-    def _destroy(self, **kwargs: Any) -> None:
+    def _destroy(self) -> None:
 
         # TODO: Determine if atomic operation is needed.
 
@@ -200,8 +194,7 @@ class SzAbstractFactory(SzAbstractFactoryAbstract):
             sz_product = SzProduct()
             sz_product._destroy()  # pylint: disable=W0212
 
-    def reinitialize(self, config_id: int, **kwargs: Any) -> None:
-        _ = kwargs
+    def reinitialize(self, config_id: int) -> None:
 
         # TODO: Determine if atomic operation is needed.