+++ src/senzing/szconfig.py
@@ -93,7 +93,7 @@ class SzConfigSaveResult(SzResponseAsCharPointerResult):
 
 class SzConfig(SzConfigAbstract):
     """
-    Use SzAbstractFactory.create_sz_config() to create an SzConfig object.
+    Use SzAbstractFactory.create_config() to create an SzConfig object.
     The SzConfig object uses the parameters provided to the SzAbstractFactory()
     function.
 
@@ -102,7 +102,7 @@ class SzConfig(SzConfigAbstract):
     .. code-block:: python
 
         sz_abstract_factory = SzAbstractFactory(instance_name, settings)
-        sz_config = sz_abstract_factory.create_sz_config()
+        sz_config = sz_abstract_factory.create_config()
 
     Parameters:
 
@@ -117,10 +117,7 @@ class SzConfig(SzConfigAbstract):
     # Python dunder/magic methods
     # -------------------------------------------------------------------------
 
-    def __init__(
-        self,
-        **kwargs: Any,
-    ) -> None:
+    def __init__(self, **kwargs: Any) -> None:
         """
         Constructor
 
@@ -189,7 +186,6 @@ class SzConfig(SzConfigAbstract):
         self,
         config_handle: int,
         data_source_code: str,
-        **kwargs: Any,
     ) -> str:
         result = self.library_handle.SzConfig_addDataSource_helper(
             as_c_uintptr_t(config_handle),
@@ -201,11 +197,11 @@ class SzConfig(SzConfigAbstract):
             return as_python_str(result.response)
 
     @catch_non_sz_exceptions
-    def close_config(self, config_handle: int, **kwargs: Any) -> None:
+    def close_config(self, config_handle: int) -> None:
         result = self.library_handle.SzConfig_close_helper(as_c_uintptr_t(config_handle))
         self.check_result(result)
 
-    def create_config(self, **kwargs: Any) -> int:
+    def create_config(self) -> int:
         result = self.library_handle.SzConfig_create_helper()
         self.check_result(result.return_code)
         return result.response  # type: ignore[no-any-return]
@@ -215,7 +211,6 @@ class SzConfig(SzConfigAbstract):
         self,
         config_handle: int,
         data_source_code: str,
-        **kwargs: Any,
     ) -> None:
         result = self.library_handle.SzConfig_deleteDataSource_helper(
             as_c_uintptr_t(config_handle),
@@ -223,18 +218,18 @@ class SzConfig(SzConfigAbstract):
         )
         self.check_result(result)
 
-    def _destroy(self, **kwargs: Any) -> None:
+    def _destroy(self) -> None:
         _ = self.library_handle.SzConfig_destroy()
 
     @catch_non_sz_exceptions
-    def export_config(self, config_handle: int, **kwargs: Any) -> str:
+    def export_config(self, config_handle: int) -> str:
         result = self.library_handle.SzConfig_save_helper(as_c_uintptr_t(config_handle))
         with FreeCResources(self.library_handle, result.response):
             self.check_result(result.return_code)
             return as_python_str(result.response)
 
     @catch_non_sz_exceptions
-    def get_data_sources(self, config_handle: int, **kwargs: Any) -> str:
+    def get_data_sources(self, config_handle: int) -> str:
         result = self.library_handle.SzConfig_listDataSources_helper(as_c_uintptr_t(config_handle))
         with FreeCResources(self.library_handle, result.response):
             self.check_result(result.return_code)
@@ -246,7 +241,6 @@ class SzConfig(SzConfigAbstract):
         instance_name: str,
         settings: Union[str, Dict[Any, Any]],
         verbose_logging: int = 0,
-        **kwargs: Any,
     ) -> None:
         result = self.library_handle.SzConfig_init(
             as_c_char_p(instance_name),
@@ -256,7 +250,7 @@ class SzConfig(SzConfigAbstract):
         self.check_result(result)
 
     @catch_non_sz_exceptions
-    def import_config(self, config_definition: str, **kwargs: Any) -> int:
+    def import_config(self, config_definition: str) -> int:
         result = self.library_handle.SzConfig_load_helper(as_c_char_p(config_definition))
         self.check_result(result.return_code)
         return result.response  # type: ignore[no-any-return]