+++ src/senzing/szproduct.py
@@ -46,7 +46,7 @@ __updated__ = "2023-11-07"
 
 class SzProduct(SzProductAbstract):
     """
-    Use SzAbstractFactory.create_sz_product() to create an SzProduct object.
+    Use SzAbstractFactory.create_product() to create an SzProduct object.
     The SzProduct object uses the parameters provided to the SzAbstractFactory()
     function.
 
@@ -55,7 +55,7 @@ class SzProduct(SzProductAbstract):
     .. code-block:: python
 
         sz_abstract_factory = SzAbstractFactory(instance_name, settings)
-        sz_product = sz_abstract_factory.create_sz_product()
+        sz_product = sz_abstract_factory.create_product()
 
     Parameters:
 
@@ -70,10 +70,7 @@ class SzProduct(SzProductAbstract):
     # Python dunder/magic methods
     # -------------------------------------------------------------------------
 
-    def __init__(
-        self,
-        **kwargs: Any,
-    ) -> None:
+    def __init__(self, **kwargs: Any) -> None:
         """
         Constructor
 
@@ -111,7 +108,7 @@ class SzProduct(SzProductAbstract):
     # SzProduct methods
     # -------------------------------------------------------------------------
 
-    def _destroy(self, **kwargs: Any) -> None:
+    def _destroy(self) -> None:
         _ = self.library_handle.SzProduct_destroy()
 
     @catch_non_sz_exceptions
@@ -120,7 +117,6 @@ class SzProduct(SzProductAbstract):
         instance_name: str,
         settings: Union[str, Dict[Any, Any]],
         verbose_logging: int = 0,
-        **kwargs: Any,
     ) -> None:
         result = self.library_handle.SzProduct_init(
             as_c_char_p(instance_name),
@@ -129,8 +125,10 @@ class SzProduct(SzProductAbstract):
         )
         self.check_result(result)
 
-    def get_license(self, **kwargs: Any) -> str:
+    def get_license(self) -> str:
         return as_python_str(self.library_handle.SzProduct_license())
 
-    def get_version(self, **kwargs: Any) -> str:
+    def get_version(
+        self,
+    ) -> str:
         return as_python_str(self.library_handle.SzProduct_version())